requirements.txt
jsonpickle
tensorflow 
numpy
flask
Pillow
